@@ -1,7 +1,7 @@
 {
   "name": "xihan-display",
-  "private": true,
-  "version": "0.0.1",
+  "private": false,
+  "version": "0.3.6",
   "type": "module",
   "description": "Xihan frontend project. A fast, lightweight, efficient, and dedicated development framework and new decentralized full scenario application software. Built based on DotNet and Vue.",
   "author": {
@@ -10,7 +10,7 @@
     "url": "https://github.com/XiHanFun/XiHan.Display"
   },
   "license": "MulanPSL-2.0",
-  "homepage": "https://github.com/XiHanFun/XiHan.Display",
+  "homepage": "https://docs.xihan.fun",
   "repository": {
     "url": "https://github.com/XiHanFun/XiHan.Display.git"
   },
@@ -43,37 +43,52 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
-    "@microsoft/signalr": "^7.0.11",
-    "axios": "^1.5.0",
-    "naive-ui": "^2.34.4",
-    "pinia": "^2.1.6",
-    "vue": "^3.3.4",
-    "vue-router": "^4.2.4"
+    "vue": "^3.3.8",
+    "vue-router": "^4.2.5",
+    "axios": "^1.6.1",
+    "pinia": "^2.1.7",
+    "naive-ui": "^2.35.0",
+    "vfonts": "^0.0.3",
+    "echarts": "^5.4.3",
+    "nprogress": "^0.2.0",
+    "qrcode": "^1.5.3",
+    "@microsoft/signalr": "^7.0.12"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.4.0",
+    "@vitejs/plugin-vue": "^4.4.1",
+    "typescript": "^5.2.2",
+    "vite": "^4.5.0",
+    "vue-tsc": "^1.8.22",
+    "@rushstack/eslint-patch": "^1.5.1",
     "@tsconfig/node18": "^18.2.2",
-    "@types/jsdom": "^21.1.2",
-    "@types/node": "^20.6.1",
+    "@types/jsdom": "^21.1.5",
+    "@types/node": "^20.9.0",
+    "@typescript-eslint/utils": "^6.11.0",
+    "@typescript-eslint/parser": "^6.11.0",
     "@vicons/ionicons4": "^0.12.0",
-    "@vitejs/plugin-vue": "^4.3.4",
     "@vue/eslint-config-prettier": "^8.0.0",
-    "@vue/eslint-config-typescript": "^12.0.0",
     "@vue/test-utils": "^2.4.1",
     "@vue/tsconfig": "^0.4.0",
-    "eslint": "^8.49.0",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint": "^8.53.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-config-standard-with-typescript": "^39.1.1",
+    "eslint-define-config": "^1.24.1",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-n": "^16.3.1",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vue": "^9.18.1",
     "jsdom": "^22.1.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^3.0.3",
-    "rimraf": "^5.0.1",
-    "typescript": "~5.2.2",
-    "unplugin-auto-import": "^0.16.6",
+    "prettier": "^3.1.0",
+    "rimraf": "^5.0.5",
+    "stylelint": "^15.11.0",
+    "stylelint-config-standard": "^34.0.0",
+    "stylelint-order": "^6.0.3",
+    "unplugin-auto-import": "^0.16.7",
     "unplugin-vue-components": "^0.25.2",
-    "vfonts": "^0.0.3",
-    "vite": "^4.4.9",
-    "vitest": "^0.34.4",
-    "vue-tsc": "^1.8.11"
+    "vitest": "^0.34.6",
+    "vue-eslint-parser": "^9.3.2"
   },
   "packageManager": "pnpm@8.1.0"
 }
