@@ -1,18 +1,29 @@
 module.exports = {
-  printWidth: 120,
+  // 超过最大值换行
+  printWidth: 100,
+  // 缩进字节数
   tabWidth: 2,
+  // 缩进不使用tab，使用空格
   useTabs: false,
+  // 句尾添加分号
   semi: true,
   vueIndentScriptAndStyle: true,
+  // 使用单引号代替双引号
   singleQuote: true,
   quoteProps: 'as-needed',
+  // 在对象，数组括号与文字之间加空格 "{ foo: bar }"
   bracketSpacing: true,
-  trailingComma: 'all',
+  // 在对象或数组最后一个元素后面是否加逗号(在ES5中加尾逗号)
+  trailingComma: 'es5',
   jsxSingleQuote: false,
-  arrowParens: 'always',
+  //  (x) => {} 箭头函数参数只有一个时是否要有小括号。avoid：省略括号
+  arrowParens: 'avoid',
   insertPragma: false,
   requirePragma: false,
-  proseWrap: 'never',
+  // 默认值。因为使用了一些折行敏感型的渲染器(如GitHub comment)而按照markdown文本样式进行折行
+  proseWrap: 'preserve',
   htmlWhitespaceSensitivity: 'strict',
-  endOfLine: 'auto',
+  // 结尾
+  endOfLine: 'lf',
+  rangeStart: 0,
 };
