@@ -7,26 +7,123 @@ export {}
 
 declare module 'vue' {
   export interface GlobalComponents {
-    NAlert: typeof import('naive-ui')['NAlert']
+    ApiCascader: typeof import('./src/components/Form/src/components/ApiCascader.vue')['default']
+    ApiRadioGroup: typeof import('./src/components/Form/src/components/ApiRadioGroup.vue')['default']
+    ApiSelect: typeof import('./src/components/Form/src/components/ApiSelect.vue')['default']
+    ApiTransfer: typeof import('./src/components/Form/src/components/ApiTransfer.vue')['default']
+    ApiTree: typeof import('./src/components/Form/src/components/ApiTree.vue')['default']
+    ApiTreeSelect: typeof import('./src/components/Form/src/components/ApiTreeSelect.vue')['default']
+    AppDarkModeToggle: typeof import('./src/components/Application/src/AppDarkModeToggle.vue')['default']
+    AppLocalePicker: typeof import('./src/components/Application/src/AppLocalePicker.vue')['default']
+    AppLogo: typeof import('./src/components/Application/src/AppLogo.vue')['default']
+    AppProvider: typeof import('./src/components/Application/src/AppProvider.vue')['default']
+    AppSearch: typeof import('./src/components/Application/src/search/AppSearch.vue')['default']
+    AppSearchFooter: typeof import('./src/components/Application/src/search/AppSearchFooter.vue')['default']
+    AppSearchKeyItem: typeof import('./src/components/Application/src/search/AppSearchKeyItem.vue')['default']
+    AppSearchModal: typeof import('./src/components/Application/src/search/AppSearchModal.vue')['default']
+    Authority: typeof import('./src/components/Authority/src/Authority.vue')['default']
+    BasicArrow: typeof import('./src/components/Basic/src/BasicArrow.vue')['default']
+    BasicButton: typeof import('./src/components/Button/src/BasicButton.vue')['default']
+    BasicDrawer: typeof import('./src/components/Drawer/src/BasicDrawer.vue')['default']
+    BasicForm: typeof import('./src/components/Form/src/BasicForm.vue')['default']
+    BasicHelp: typeof import('./src/components/Basic/src/BasicHelp.vue')['default']
+    BasicMenu: typeof import('./src/components/Menu/src/BasicMenu.vue')['default']
+    BasicMenuItem: typeof import('./src/components/Menu/src/components/BasicMenuItem.vue')['default']
+    BasicModal: typeof import('./src/components/Modal/src/BasicModal.vue')['default']
+    BasicSubMenuItem: typeof import('./src/components/Menu/src/components/BasicSubMenuItem.vue')['default']
+    BasicTable: typeof import('./src/components/Table/src/BasicTable.vue')['default']
+    BasicTitle: typeof import('./src/components/Basic/src/BasicTitle.vue')['default']
+    BasicTree: typeof import('./src/components/Tree/src/BasicTree.vue')['default']
+    BasicUpload: typeof import('./src/components/Upload/src/BasicUpload.vue')['default']
+    CardList: typeof import('./src/components/CardList/src/CardList.vue')['default']
+    ClickOutSide: typeof import('./src/components/ClickOutSide/src/ClickOutSide.vue')['default']
+    CodeEditor: typeof import('./src/components/CodeEditor/src/CodeEditor.vue')['default']
+    CodeMirror: typeof import('./src/components/CodeEditor/src/codemirror/CodeMirror.vue')['default']
+    CollapseContainer: typeof import('./src/components/Container/src/collapse/CollapseContainer.vue')['default']
+    CollapseHeader: typeof import('./src/components/Container/src/collapse/CollapseHeader.vue')['default']
+    CollapseTransition: typeof import('./src/components/Transition/src/CollapseTransition.vue')['default']
+    ColumnSetting: typeof import('./src/components/Table/src/components/settings/ColumnSetting.vue')['default']
+    ContextMenu: typeof import('./src/components/ContextMenu/src/ContextMenu.vue')['default']
+    CountButton: typeof import('./src/components/CountDown/src/CountButton.vue')['default']
+    CountdownInput: typeof import('./src/components/CountDown/src/CountdownInput.vue')['default']
+    CountTo: typeof import('./src/components/CountTo/src/CountTo.vue')['default']
+    Cropper: typeof import('./src/components/Cropper/src/Cropper.vue')['default']
+    CropperAvatar: typeof import('./src/components/Cropper/src/CropperAvatar.vue')['default']
+    CropperModal: typeof import('./src/components/Cropper/src/CropperModal.vue')['default']
+    Description: typeof import('./src/components/Description/src/Description.vue')['default']
+    Dialog: typeof import('./src/components/Prompt/dialog.vue')['default']
+    DragVerify: typeof import('./src/components/Verify/src/DragVerify.vue')['default']
+    DrawerFooter: typeof import('./src/components/Drawer/src/components/DrawerFooter.vue')['default']
+    DrawerHeader: typeof import('./src/components/Drawer/src/components/DrawerHeader.vue')['default']
+    Dropdown: typeof import('./src/components/Dropdown/src/Dropdown.vue')['default']
+    EditableCell: typeof import('./src/components/Table/src/components/editable/EditableCell.vue')['default']
+    Editor: typeof import('./src/components/Tinymce/src/Editor.vue')['default']
+    EditTableHeaderIcon: typeof import('./src/components/Table/src/components/EditTableHeaderIcon.vue')['default']
+    EllipsisText: typeof import('./src/components/EllipsisText/src/EllipsisText.vue')['default']
+    ExportExcelModal: typeof import('./src/components/Excel/src/ExportExcelModal.vue')['default']
+    FileList: typeof import('./src/components/Upload/src/components/FileList.vue')['default']
+    FlowChart: typeof import('./src/components/FlowChart/src/FlowChart.vue')['default']
+    FlowChartToolbar: typeof import('./src/components/FlowChart/src/FlowChartToolbar.vue')['default']
+    FormAction: typeof import('./src/components/Form/src/components/FormAction.vue')['default']
+    FormItem: typeof import('./src/components/Form/src/components/FormItem.vue')['default']
+    FullScreenSetting: typeof import('./src/components/Table/src/components/settings/FullScreenSetting.vue')['default']
+    Functional: typeof import('./src/components/Preview/src/Functional.vue')['default']
+    HeaderCell: typeof import('./src/components/Table/src/components/HeaderCell.vue')['default']
+    Icon: typeof import('./src/components/Icon/Icon.vue')['default']
+    IconPicker: typeof import('./src/components/Icon/src/IconPicker.vue')['default']
+    ImageUpload: typeof import('./src/components/Upload/src/components/ImageUpload.vue')['default']
+    ImgRotate: typeof import('./src/components/Verify/src/ImgRotate.vue')['default']
+    ImgUpload: typeof import('./src/components/Tinymce/src/ImgUpload.vue')['default']
+    ImportExcel: typeof import('./src/components/Excel/src/ImportExcel.vue')['default']
+    JsonPreview: typeof import('./src/components/CodeEditor/src/json-preview/JsonPreview.vue')['default']
+    Loading: typeof import('./src/components/Loading/src/Loading.vue')['default']
+    Markdown: typeof import('./src/components/Markdown/src/Markdown.vue')['default']
+    MarkdownViewer: typeof import('./src/components/Markdown/src/MarkdownViewer.vue')['default']
+    Menu: typeof import('./src/components/SimpleMenu/src/components/Menu.vue')['default']
+    MenuCollapseTransition: typeof import('./src/components/SimpleMenu/src/components/MenuCollapseTransition.vue')['default']
+    MenuItem: typeof import('./src/components/SimpleMenu/src/components/MenuItem.vue')['default']
+    MenuItemContent: typeof import('./src/components/Menu/src/components/MenuItemContent.vue')['default']
+    ModalClose: typeof import('./src/components/Modal/src/components/ModalClose.vue')['default']
+    ModalFooter: typeof import('./src/components/Modal/src/components/ModalFooter.vue')['default']
+    ModalHeader: typeof import('./src/components/Modal/src/components/ModalHeader.vue')['default']
+    ModalWrapper: typeof import('./src/components/Modal/src/components/ModalWrapper.vue')['default']
     NAvatar: typeof import('naive-ui')['NAvatar']
     NBadge: typeof import('naive-ui')['NBadge']
     NButton: typeof import('naive-ui')['NButton']
     NButtonGroup: typeof import('naive-ui')['NButtonGroup']
     NIcon: typeof import('naive-ui')['NIcon']
-    NLayout: typeof import('naive-ui')['NLayout']
-    NLayoutSider: typeof import('naive-ui')['NLayoutSider']
-    NMenu: typeof import('naive-ui')['NMenu']
-    NProgress: typeof import('naive-ui')['NProgress']
-    NRadioButton: typeof import('naive-ui')['NRadioButton']
-    NRadioGroup: typeof import('naive-ui')['NRadioGroup']
-    NSkeleton: typeof import('naive-ui')['NSkeleton']
     NSpace: typeof import('naive-ui')['NSpace']
-    NStep: typeof import('naive-ui')['NStep']
-    NSteps: typeof import('naive-ui')['NSteps']
     NSwitch: typeof import('naive-ui')['NSwitch']
-    NTimeline: typeof import('naive-ui')['NTimeline']
-    NTimelineItem: typeof import('naive-ui')['NTimelineItem']
+    PageFooter: typeof import('./src/components/Page/src/PageFooter.vue')['default']
+    PageWrapper: typeof import('./src/components/Page/src/PageWrapper.vue')['default']
+    PopConfirmButton: typeof import('./src/components/Button/src/PopConfirmButton.vue')['default']
+    Preview: typeof import('./src/components/Preview/src/Preview.vue')['default']
+    Qrcode: typeof import('./src/components/Qrcode/src/Qrcode.vue')['default']
+    RadioButtonGroup: typeof import('./src/components/Form/src/components/RadioButtonGroup.vue')['default']
+    RedoSetting: typeof import('./src/components/Table/src/components/settings/RedoSetting.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
+    Scrollbar: typeof import('./src/components/Scrollbar/src/Scrollbar.vue')['default']
+    ScrollContainer: typeof import('./src/components/Container/src/ScrollContainer.vue')['default']
+    Settings: typeof import('./src/components/Table/src/components/settings/index.vue')['default']
+    SimpleMenu: typeof import('./src/components/SimpleMenu/src/SimpleMenu.vue')['default']
+    SimpleMenuTag: typeof import('./src/components/SimpleMenu/src/SimpleMenuTag.vue')['default']
+    SimpleSubMenu: typeof import('./src/components/SimpleMenu/src/SimpleSubMenu.vue')['default']
+    SizeSetting: typeof import('./src/components/Table/src/components/settings/SizeSetting.vue')['default']
+    StrengthMeter: typeof import('./src/components/StrengthMeter/src/StrengthMeter.vue')['default']
+    SubMenuItem: typeof import('./src/components/SimpleMenu/src/components/SubMenuItem.vue')['default']
+    SvgIcon: typeof import('./src/components/Icon/src/SvgIcon.vue')['default']
+    TableAction: typeof import('./src/components/Table/src/components/TableAction.vue')['default']
+    TableFooter: typeof import('./src/components/Table/src/components/TableFooter.vue')['default']
+    TableHeader: typeof import('./src/components/Table/src/components/TableHeader.vue')['default']
+    TableImg: typeof import('./src/components/Table/src/components/TableImg.vue')['default']
+    TableTitle: typeof import('./src/components/Table/src/components/TableTitle.vue')['default']
+    ThumbUrl: typeof import('./src/components/Upload/src/components/ThumbUrl.vue')['default']
+    Time: typeof import('./src/components/Time/src/Time.vue')['default']
+    Tooltip: typeof import('./src/components/EllipsisText/src/Tooltip.vue')['default']
+    TreeHeader: typeof import('./src/components/Tree/src/components/TreeHeader.vue')['default']
+    UploadModal: typeof import('./src/components/Upload/src/components/UploadModal.vue')['default']
+    UploadPreviewModal: typeof import('./src/components/Upload/src/components/UploadPreviewModal.vue')['default']
+    VirtualScroll: typeof import('./src/components/VirtualScroll/src/VirtualScroll.vue')['default']
   }
 }
